# docs/spikes/S7-api-v1/recorded/swagger_v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "MintPlayer API",
    "description": "API that allows access to the data of MintPlayer. ",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/Tag/page": {
      "post": {
        "tags": [
          "Tag"
        ],
        "summary": "Page through the tags in the database.",
        "operationId": "api-tag-page",
        "requestBody": {
          "description": "Object containing the pagination information.",
          "content": {
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/TagPaginationRequest"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/TagPaginationRequest"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/TagPaginationRequest"
              }
            },
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/TagPaginationRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagPaginationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TagPaginationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TagPaginationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/TagPaginationResponse"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/TagPaginationResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagPaginationResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagPaginationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/Tag/{id}": {
      "get": {
        "tags": [
          "Tag"
        ],
        "summary": "Get a specific tag from the database.",
        "operationId": "api-tag-get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the tag to retrieve.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "include_relations",
            "in": "header",
            "description": "Specifies whether the related entities should be included in the response.",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Tag"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Tag"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tag"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tag"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Tag"
        ],
        "summary": "Updates a tag in the database.",
        "operationId": "api-tag-update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "New tag information.",
          "content": {
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Tag"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/Tag"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/Tag"
              }
            },
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Tag"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Tag"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Tag"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Tag"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Tag"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Tag"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tag"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tag"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tag"
        ],
        "summary": "Deletes a tag from the database.",
        "operationId": "api-tag-delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the tag to delete.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/TagCategory/page": {
      "post": {
        "tags": [
          "TagCategory"
        ],
        "summary": "Page through the tag categories in the database.",
        "operationId": "api-tagcategory-page",
        "requestBody": {
          "description": "Object containing the pagination information.",
          "content": {
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/TagCategoryPaginationRequest"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/TagCategoryPaginationRequest"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/TagCategoryPaginationRequest"
              }
            },
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/TagCategoryPaginationRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagCategoryPaginationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TagCategoryPaginationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TagCategoryPaginationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/TagCategoryPaginationResponse"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/TagCategoryPaginationResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagCategoryPaginationResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagCategoryPaginationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/TagCategory/{id}": {
      "get": {
        "tags": [
          "TagCategory"
        ],
        "summary": "Get a specific tag category from the database.",
        "operationId": "api-tagcategory-get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the tag category to retrieve.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "include_relations",
            "in": "header",
            "description": "Specifies whether the related entities should be included in the response.",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/TagCategory"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/TagCategory"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagCategory"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagCategory"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "TagCategory"
        ],
        "summary": "Updates a tag category in the database.",
        "operationId": "api-tagcategory-update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "New tag category information.",
          "content": {
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/TagCategory"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/TagCategory"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/TagCategory"
              }
            },
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/TagCategory"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagCategory"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TagCategory"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TagCategory"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/TagCategory"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/TagCategory"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagCategory"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagCategory"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "TagCategory"
        ],
        "summary": "Deletes a tag category from the database.",
        "operationId": "api-tagcategory-delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the tag category to delete.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Color": {
        "type": "object",
        "properties": {
          "r": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "g": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "b": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "a": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "isKnownColor": {
            "type": "boolean",
            "readOnly": true
          },
          "isEmpty": {
            "type": "boolean",
            "readOnly": true
          },
          "isNamedColor": {
            "type": "boolean",
            "readOnly": true
          },
          "isSystemColor": {
            "type": "boolean",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "ListSortDirection": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "Medium": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "$ref": "#/components/schemas/MediumType"
          },
          "value": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MediumType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "visible": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "SortColumn": {
        "type": "object",
        "properties": {
          "property": {
            "type": "string",
            "nullable": true
          },
          "direction": {
            "$ref": "#/components/schemas/SortDirection"
          }
        },
        "additionalProperties": false
      },
      "SortDirection": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
      "Subject": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "media": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Medium"
            },
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tag"
            },
            "nullable": true
          },
          "text": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "concurrencyStamp": {
            "type": "string",
            "nullable": true
          },
          "dateUpdate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "Tag": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "$ref": "#/components/schemas/TagCategory"
          },
          "subjects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Subject"
            },
            "nullable": true
          },
          "parent": {
            "$ref": "#/components/schemas/Tag"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tag"
            },
            "nullable": true
          },
          "text": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "TagCategory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "color": {
            "$ref": "#/components/schemas/Color"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tag"
            },
            "nullable": true
          },
          "totalTagCount": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "TagCategoryPaginationRequest": {
        "type": "object",
        "properties": {
          "perPage": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "sortColumns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortColumn"
            },
            "nullable": true
          },
          "sortProperty": {
            "type": "string",
            "nullable": true
          },
          "sortDirection": {
            "$ref": "#/components/schemas/ListSortDirection"
          }
        },
        "additionalProperties": false
      },
      "TagCategoryPaginationResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagCategory"
            },
            "nullable": true
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "perPage": {
            "type": "integer",
            "format": "int32"
          },
          "totalRecords": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "TagPaginationRequest": {
        "type": "object",
        "properties": {
          "perPage": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "sortColumns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortColumn"
            },
            "nullable": true
          },
          "sortProperty": {
            "type": "string",
            "nullable": true
          },
          "sortDirection": {
            "$ref": "#/components/schemas/ListSortDirection"
          }
        },
        "additionalProperties": false
      },
      "TagPaginationResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tag"
            },
            "nullable": true
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "perPage": {
            "type": "integer",
            "format": "int32"
          },
          "totalRecords": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}